cli support register dataset
